move blog template to it's own file
